requirements.txt
Django~=5.1.4
python-dotenv~=1.0.1
psycopg~=3.2.3
django-jazzmin==3.0.1
PyJWT~=2.10.1
